optimized structure
following a much simpler folder and package structure that containing most of the codes in file core.py
historical data in module history.py
kept other
