+++ qteasy/__main__.py
@@ -1,5 +1,7 @@
 from qteasy import app
 
+
+
 if __name__ == '__main__':
     app.run()
     app.stop()
