--- qteasy/app.py
@@ -1,10 +1,0 @@
-def run():
-    print('function \'run\' has been called')
-    return None
-
-def stop():
-    print('function \'stop\' has been called')
-    return None
-
-def activate():
-    print{'function \'activated\' from module app has been called!'}
