--- qteasy/app/strategy/__init__.py
@@ -1,1 +1,0 @@
-
